Split up touch.rs into smaller chunks
