+++ crates/bevy_input/src/touch/mod.rs
@@ -1,0 +1,7 @@
+pub mod event;
+pub mod force;
+pub mod phase;
+pub mod system;
+pub mod touches;
+
+pub use crate::touch::{event::*, force::*, phase::*, system::*, touches::*};
